Added a class to handle FFTs and removed FFTW support in the fftconv_p mex file
This a hack (with lots of #ifdef:s) to avoid crashes when compting FFTs in mex-files. It
looks like there is a conflict bewteen the FFTW lib and the FFTW lib version bundeled
with Matlab. Here we use mexCallMatlab to compute FFTs since we cannot use the FFTW API
but calling mexCallMatlab does not seem to be thread safe so we havle also disabled all
threading for the fftconv_p mex-file.