Made the dreamcylind mex/oct-file use the Attenuation class (with attenuation threading disabled in the mex-file)
This function replaces the two former dreamdylind_f and dreamcylind_d functions where we now
use the curvature parameter to specify if it is a focused/concave transducer or a defocused/convex
transducer. Also, deactivated some code that most likely protects from using observation points
inside the transducer or from points that cannot see the surface of the transducer. However,
this code needs to be revised before we can activate it again.