Made the dream_arr_cylind mex/oct-file use the Attenuation class (with attenuation threading disabled in the mex-file)
This function replaces the two former dream_arr_cylind_f and dream_arr_cylind_d functions where we now
use the curvature parameter to specify if it is a focused/concave transducer or a defocused/convex
in the same way as the dreamcylind mex/oct-file now also do. We also re-use the dreamcylind function to reduce
code duplication.