Made the fftconv_p mex- and oct-files use (almost) the same code and forced FFTW
We have now enabled the mex-files to link to a shared FFTW lib which
is problementic due to conflicts between libmwfftw3 and (system) libfftw3.
The current work-a-round on Linux is to pre-load the libfftw3.so using
something like LD_PRELOAD=/usr/lib64/libfftw3.so

We have also made the fftconv_p mex- and oct-files very similar now.