Made DAS objects persist in the das oct- and mex-files
Now the das init (constructor) only runs the first time we call the das oct/mex-function
if the das operation is the same (same buffer size and das method). This means
that we only run the OpenCL buffer allocations and init once if we do repeated
das operation with the same setup.