Inital delay-and-sum (DAS) processing python support #1017
This adds initial single and double DAS processing support for python. Both
CPU and GPU processing (for SAFT, TFM, and RCA) is supported. There is no
tests yet since we need fftconv_p to make the simulated data which is
not implemented for python yet. To support both single and double precision
we templated the ArgParser class (we can also use this for supporting single
precision for the transducer functions later).